requirements.txt
# Core data
pandas==2.2.2
numpy==1.26.4
sqlalchemy==2.0.30
psycopg2-binary==2.9.9

# Data cleaning
rapidfuzz==3.9.3
regex==2024.5.15

# HTTP / API
requests==2.32.3
tqdm==4.66.4
kaggle==1.7.4.5

# Dashboard
streamlit==1.35.0
plotly==5.22.0

# Stats
scipy==1.13.1

# Dev / testing
pytest==8.2.2
python-dotenv==1.0.1
jupyter==1.0.0
ipykernel==6.29.4
